Order cancellations, refunds and emails...

+++ desapp-groupj-backend/src/test/java/ar/edu/unq/desapp/groupj/backend/model/OrderDetailBuilderTest.java
@@ -15,6 +15,6 @@ public class OrderDetailBuilderTest {
                 withRequestedAmount(1).
                 build();
 
-        assertEquals(1, orderDetail.getRequestedAmount());
+        assertEquals((Integer)1, orderDetail.getRequestedAmount());
     }
 }
